[SPIR-V] Add structurizer pass
This commit adds the initial structurizer bits. This pass uses the
convergence tokens to determine convergence regions, and emit an
OpLoopMerge when required.

This structurizer is far from complere: only SESE regions are supported,
no OpSelectionMerge are emited. But this first step allows us to support
an initial set of HLSL shaders.

Signed-off-by: Nathan Gauër <brioche@google.com>