[SPIR-V] Add SPIR-V structurizer
This commit  adds an initial SPIR-V structurizer.
It leverages the previously merged passes, and the convergence region
analysis to determine the correct merge and continue blocks for SPIR-V.

The first part replaces switches with a single edge with a simple
branch.
Then, we add OpLoopMerge instruction to each loop encountered.
Then, we add OpSelectionMerge for each conditional branch which has
no OpLoopMerge instruction.

Finally, we fixup the merge instructions:
 - we split blocks with multiple headers into 2 blocks.
 - we split blocks that are a merge blocks for 2 or more constructs:
        SPIR-V spec disallow a merge block to be shared by 2
        loop/switch/condition construct.
 - we split merge & continue blocks:
        SPIR-V spec disallow a basic block to be both a continue
        block, and a merge block.
 - we remove superfluous headers: when a header doesn't bring more info
   than the parent on the divergence state, it must be removed.
 - We sort blocks depending on the dominator tree order:
        SPIR-V spec requires blocks to be sorted in a specific way.

As is, this code seems to pass most DXC structurization tests, except
the ones relying on unrelated features this backend doesn't yet support
like i64 switches, boolean types.
One known case fails, but it is because the MergeExit pass doesn't
supports switches yet (there is a FIXME).
This PR is already big-enough as-is, so I think we should get this in
first, and then add support for those switches.

Signed-off-by: Nathan Gauër <brioche@google.com>