[IR] Prevent DCE from removing convergence intrinsics
When a callee is marked as `convergent`, some targets like HLSL/SPIR-V
add a convergent token to the call.
This is valid if both functions are marked as `convergent`.

ADCE/BDCE and other DCE passes were allowed to remove convergence
intrinsics when their token were unused.
This meant a leaf function could lose all its convergence intrinsics.
This would allow further optimization to remove the `convergent`
attribute from the callee.
Issue was the caller was not updated, and we now had a convergence token
attached to a call function calling a non-convergent function.

This can be solved in different ways:
 - modify DCE passes to keep calls marked as convergent.
 - add a 'hasSideEffect' to the intrinsics definition, preventing DCE to
   remove the intrinsic.
 - fix optimization passed removing the attribute by also removing the
   tokens from the call.

- Widening the definition of `hasSideEffect` is not right.
Leaving us with 1 and 3.
- Fixing callers would require to transform a function pass into a
  module pass. A larger change, and propagating the attribute change
  could cause other issues.

For this reason, I think the simplest/best option is 1: prevent DCE from
removing convergence intrinsics.